+++ packages/terra-form-select/src/_NoResults.jsx
@@ -1,11 +1,17 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 import classNames from 'classnames/bind';
+import { injectIntl, intlShape } from 'react-intl';
 import styles from './_NoResults.module.scss';
 
 const cx = classNames.bind(styles);
 
 const propTypes = {
+  /**
+   * @private
+   * The intl object containing translations. This is retrieved from the context automatically by injectIntl.
+   */
+  intl: intlShape.isRequired,
   /**
    * Content to display when no results are found.
    */
@@ -16,22 +22,12 @@ const propTypes = {
   value: PropTypes.oneOfType([PropTypes.number, PropTypes.string]),
 };
 
-const contextTypes = {
-  /* eslint-disable consistent-return */
-  intl: (context) => {
-    if (context.intl === undefined) {
-      return new Error('Component is internationalized, and must be wrapped in terra-base');
-    }
-  },
-};
-
-const NoResults = ({ noResultContent, value }, context) => (
+const NoResults = ({ intl, noResultContent, value }) => (
   <div role="option" className={cx('no-results')} aria-selected="false">
-    {noResultContent || context.intl.formatMessage({ id: 'Terra.form.select.noResults' }, { text: value })}
+    {noResultContent || intl.formatMessage({ id: 'Terra.form.select.noResults' }, { text: value })}
   </div>
 );
 
 NoResults.propTypes = propTypes;
-NoResults.contextTypes = contextTypes;
 
-export default NoResults;
+export default injectIntl(NoResults);